Import and Format MIMIC-IV

Import and format MIMIC-IV dataset as done in the AMPEL project.

import_mimic(path, verbose = interactive())

Arguments

path

character(1), path to the root MIMIC-IV folder (that contains the subfolders: core and hosp).

verbose

logical(1), if TRUE progress messages are shown.

Value

data.frame, same as x but with an added column Excluded and 3 attributes "message", "n_cases", "n_cbc" that describe the processing and the kept cases and complete blood counts (CBC)

Details

For the MIMIC-IV dataset the files core/patients.csv, core/transfers.csv, hosp/labevents.csv, and hosp/diagnoses_icd.csv are needed.

References

Johnson, A., Bulgarelli, L., Pollard, T., Horng, S., Celi, L. A., & Mark, R. (2021). MIMIC-IV (version 1.0). PhysioNet. doi: 10.13026/s6n6-xd98 .

Documentation https://mimic.mit.edu/docs/iv/

Author

Daniel Steinbach daniel.steinbach@medizin.uni-leipzig.de and Sebastian Gibb mail@sebastiangibb.de

Examples

if (FALSE) {
mimic <- import_mimic("path/to/mimic-iv-1.0")
}