sbcdata: Laboratory Diagnostics from Septic and Non-septic Patients Used in the AMPEL Project

The data were collected as part of the AMPEL project (Analysis and Reporting System for the Improvement of Patient Safety through Real-Time Integration of Laboratory Findings, http://www.ampel-cdss.com/). They include administration data, sepsis labels based on ICD10 codes and laboratory diagnostics from patients admitted to the University Hospital Leipzig between Januar 2014 and December 2021 and to the University Hospital Greifswald between Januar 2015 and December 2020, respectively. Additionally it provides functions to generate a similar dataset from MIMIC-IV (Johnson et al. 2021 doi:10.13026/S6N6-XD98 ).

See also

Author

Maintainer: Sebastian Gibb mail@sebastiangibb.de (ORCID)

Authors:

Paul Ahrens paul.ahrens@medizin.uni-leipzig.de [data contributor]

Daniel Steinbach daniel.steinbach@medizin.uni-leipzig.de [data contributor]

Maria Schmidt maria.schmidt@medizin.uni-leipzig.de [data contributor]

Thorsten Kaiser thorsten.kaiser@medizin.uni-leipzig.de [data contributor]

Other contributors:

Mark Wernsdorfer mark.wernsdorfer@medizin.uni-leipzig.de [data contributor]

Matthias Nauck matthias.nauck@med.uni-greifswald.de [data contributor]

Stefan Bollmann stefan.bollmann@med.uni-greifswald.de [data contributor]

Thomas Hildebrandt thomas.hildebrandt@med.uni-greifswald.de [data contributor]